third_party/blink/web_tests/platform/generic/fast/css/dynamic-class-backdrop-pseudo-expected.txt

Change class affecting ::backdrop style

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(dialog, '::backdrop').backgroundColor is not green
PASS internals.updateStyleAndReturnAffectedElementCount() is 2
PASS getComputedStyle(dialog, '::backdrop').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
